shapeshiftr: Analyze Group/Home Range Shape Shifts Using Location Data.

R package for detecting population range reconfiguration via inter-individual distances

Details

The package is designed for spatial and temporal analysis of location data, especially those multiple sampling sessions (seasons, years, sites, populations, etc).

Reference

For the full description of the method and its application, see: Love & Otto (2025). Spatial Signatures of Population Cohesion. Methods in Ecology and Evolution. https://doi.org/##

See also

Author

Maintainer: Nicola Love nicolalove@zoology.ubc.ca